+++ widgets/common/src/Tenants/index.ts
@@ -1,0 +1,5 @@
+import TenantsDropdown from './TenantsDropdown';
+
+export default {
+    TenantsDropdown
+};
