+++ widgets/common/src/index.ts
@@ -19,6 +19,7 @@ import StagePropTypes from './props';
 import Roles from './roles';
 import SecretActions from './secrets/SecretActions';
 import TerraformModal from './terraformModal';
+import tenants from './Tenants';
 import EventUtils from './utils/EventUtils';
 
 const StageCommon = {
@@ -40,6 +41,7 @@ const StageCommon = {
     Roles,
     SecretActions,
     TerraformModal,
+    tenants,
     EventUtils,
     Consts
 };
